python/sources/csv_import/requirements.txt
fastapi[standard]
quixstreams
python-dotenv
python-multipart
